python/requirements.txt
supabase==2.3.0
python-dotenv==1.0.0
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
selenium==4.25.0

